shasha-chatbot/requirements.txt
Flask==2.2.5
gunicorn==20.1.0
pymongo[srv]==4.4.0
dnspython==2.3.0
python-dotenv==1.0.0
